+++ src/Common/Commands.Common/Extensions/CmdletExtensions.cs
@@ -210,7 +210,7 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
         /// <param name="cmdlet">the executing cmdlet</param>
         /// <param name="parameterName">The name of the parameter to return</param>
         /// <returns>true if the parameter was provided by the user, otherwise false</returns>
-        public static bool IsBound(this PSCmdlet cmdlet, string parameterName) 
+        public static bool IsBound(this PSCmdlet cmdlet, string parameterName)
         {
             return cmdlet.MyInvocation.BoundParameters.ContainsKey(parameterName);
         }
@@ -225,6 +225,10 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
             try
             {
                 return psCmdlet.ResolvePath(path);
+            }
+            catch (Exception ex) when (ex is ItemNotFoundException)
+            {
+                return Path.GetFullPath(Path.Combine(psCmdlet.SessionState.Path.CurrentLocation.Path, path));
             }
             catch
             {